First pass at initial refactor
