+++ packages/web/components/Dashboard/Filters/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Filters'
